+++ R/command-package.R
@@ -7,6 +7,7 @@
 #' and therefore more transparent, flexible, and reliable.
 #'
 #' - [cmd_assign()] Process command line arguments
+#' - [use_renv()] Activate a renv project before loading packages
 #' - [extract_shell()] Turn a `cmd_assign()` call into a shell command
 #' - [extract_make()] Turn a `cmd_assign()` call into a Makefile rule
 #' - [shell_script()] Create a shell script
@@ -15,6 +16,8 @@
 #'   How to use `cmd_assign()`
 #' - [Modular Workflows for Data Analysis](https://bayesiandemography.github.io/command/articles/workflow.html)
 #'   Safe, flexible data analysis workflows
+#' - [Using command with renv](https://bayesiandemography.github.io/command/articles/a5_renv.html)
+#'   Running pipeline scripts with renv
 #' 
 #' @docType package
 #' @name command-package